requirements.txt
pyodbc==5.2.0
python-dotenv==1.0.1
pandas==2.2.3 
tqdm==4.66.5